+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed:
+- Refactor: Combine plugins in one file, add plugins to the class docs dynamically.
 
 ## [1.0.3] - 2025-10-30
 ### Added: